clip/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
idna==3.10
requests==2.32.3
urllib3==2.4.0
youtube-transcript-api==1.0.3
yt-dlp==2025.4.30
fastapi>=0.68.0
uvicorn>=0.15.0
ffmpeg-python>=0.2.0
pydantic>=1.8.0